refactor: remove all sub modules of proposal actions module
