+++ src/context/context.module.ts
@@ -8,8 +8,6 @@ import { GroupRolesModule } from '../groups/group-roles/group-roles.module';
 import { GroupsModule } from '../groups/groups.module';
 import { ImagesModule } from '../images/images.module';
 import { PostsModule } from '../posts/posts.module';
-import { ProposalActionEventsModule } from '../proposals/proposal-actions/proposal-action-events/proposal-action-events.module';
-import { ProposalActionRolesModule } from '../proposals/proposal-actions/proposal-action-roles/proposal-action-roles.module';
 import { ProposalActionsModule } from '../proposals/proposal-actions/proposal-actions.module';
 import { ProposalsModule } from '../proposals/proposals.module';
 import { ShieldModule } from '../shield/shield.module';
@@ -27,8 +25,6 @@ import { ContextService } from './context.service';
     GroupsModule,
     ImagesModule,
     PostsModule,
-    ProposalActionEventsModule,
-    ProposalActionRolesModule,
     ProposalActionsModule,
     ProposalsModule,
     ShieldModule,
