+++ src/context/context.service.ts
@@ -8,8 +8,6 @@ import { GroupRolesService } from '../groups/group-roles/group-roles.service';
 import { GroupsService } from '../groups/groups.service';
 import { ImagesService } from '../images/images.service';
 import { PostsService } from '../posts/posts.service';
-import { ProposalActionEventsService } from '../proposals/proposal-actions/proposal-action-events/proposal-action-events.service';
-import { ProposalActionRolesService } from '../proposals/proposal-actions/proposal-action-roles/proposal-action-roles.service';
 import { ProposalActionsService } from '../proposals/proposal-actions/proposal-actions.service';
 import { ProposalsService } from '../proposals/proposals.service';
 import { ShieldService } from '../shield/shield.service';
@@ -30,8 +28,6 @@ export class ContextService {
     private groupsService: GroupsService,
     private imagesService: ImagesService,
     private postsService: PostsService,
-    private proposalActionEventsService: ProposalActionEventsService,
-    private proposalActionRolesService: ProposalActionRolesService,
     private proposalActionsService: ProposalActionsService,
     private proposalsService: ProposalsService,
     private shieldService: ShieldService,
@@ -56,8 +52,6 @@ export class ContextService {
       groupsService: this.groupsService,
       imagesService: this.imagesService,
       postsService: this.postsService,
-      proposalActionEventsService: this.proposalActionEventsService,
-      proposalActionRolesService: this.proposalActionRolesService,
       proposalActionsService: this.proposalActionsService,
       proposalsService: this.proposalsService,
       shieldService: this.shieldService,
