+++ src/context/context.types.ts
@@ -8,8 +8,6 @@ import { GroupRolesService } from '../groups/group-roles/group-roles.service';
 import { GroupsService } from '../groups/groups.service';
 import { ImagesService } from '../images/images.service';
 import { PostsService } from '../posts/posts.service';
-import { ProposalActionEventsService } from '../proposals/proposal-actions/proposal-action-events/proposal-action-events.service';
-import { ProposalActionRolesService } from '../proposals/proposal-actions/proposal-action-roles/proposal-action-roles.service';
 import { ProposalActionsService } from '../proposals/proposal-actions/proposal-actions.service';
 import { ProposalsService } from '../proposals/proposals.service';
 import { ShieldService } from '../shield/shield.service';
@@ -25,8 +23,6 @@ export interface ContextServices {
   groupsService: GroupsService;
   imagesService: ImagesService;
   postsService: PostsService;
-  proposalActionEventsService: ProposalActionEventsService;
-  proposalActionRolesService: ProposalActionRolesService;
   proposalActionsService: ProposalActionsService;
   proposalsService: ProposalsService;
   shieldService: ShieldService;
