+++ src/database/data-source.ts
@@ -15,12 +15,12 @@ import { Post } from '../posts/models/post.model';
 import { ProposalConfig } from '../proposals/models/proposal-config.model';
 import { Proposal } from '../proposals/models/proposal.model';
 import { ProposalAction } from '../proposals/proposal-actions/models/proposal-action.model';
-import { ProposalActionEventHost } from '../proposals/proposal-actions/proposal-action-events/models/proposal-action-event-host.model';
-import { ProposalActionEvent } from '../proposals/proposal-actions/proposal-action-events/models/proposal-action-event.model';
-import { ProposalActionGroupConfig } from '../proposals/proposal-actions/proposal-action-group-configs/models/proposal-action-group-config.model';
-import { ProposalActionPermission } from '../proposals/proposal-actions/proposal-action-roles/models/proposal-action-permission.model';
-import { ProposalActionRoleMember } from '../proposals/proposal-actions/proposal-action-roles/models/proposal-action-role-member.model';
-import { ProposalActionRole } from '../proposals/proposal-actions/proposal-action-roles/models/proposal-action-role.model';
+import { ProposalActionEventHost } from '../proposals/proposal-actions/models/proposal-action-event-host.model';
+import { ProposalActionEvent } from '../proposals/proposal-actions/models/proposal-action-event.model';
+import { ProposalActionGroupConfig } from '../proposals/proposal-actions/models/proposal-action-group-config.model';
+import { ProposalActionPermission } from '../proposals/proposal-actions/models/proposal-action-permission.model';
+import { ProposalActionRoleMember } from '../proposals/proposal-actions/models/proposal-action-role-member.model';
+import { ProposalActionRole } from '../proposals/proposal-actions/models/proposal-action-role.model';
 import { ServerConfig } from '../server-configs/models/server-configs.model';
 import { ServerInvite } from '../server-invites/models/server-invite.model';
 import { ServerRolePermission } from '../server-roles/models/server-role-permission.model';
