+++ src/proposals/proposal-actions/models/proposal-action-group-config.input.ts
