+++ src/proposals/proposal-actions/models/proposal-action-permission.model.ts
