+++ src/proposals/proposal-actions/models/proposal-action-role-input.ts
@@ -1,5 +1,5 @@
 import { Field, InputType, Int } from '@nestjs/graphql';
-import { GroupRolePermissionInput } from '../../../../groups/group-roles/models/group-role-permission.input';
+import { GroupRolePermissionInput } from '../../../groups/group-roles/models/group-role-permission.input';
 import { ProposalActionRoleMemberInput } from './proposal-action-role-member.input';
 
 @InputType()
