+++ src/proposals/proposal-actions/proposal-action-event-hosts.resolver.ts
@@ -1,16 +1,14 @@
 import { Parent, ResolveField, Resolver } from '@nestjs/graphql';
-import { User } from '../../../users/models/user.model';
+import { User } from '../../users/models/user.model';
 import { ProposalActionEventHost } from './models/proposal-action-event-host.model';
-import { ProposalActionEventsService } from './proposal-action-events.service';
+import { ProposalActionsService } from './proposal-actions.service';
 
 @Resolver(() => ProposalActionEventHost)
 export class ProposalActionEventHostsResolver {
-  constructor(
-    private proposalActionEventsService: ProposalActionEventsService,
-  ) {}
+  constructor(private proposalActionsService: ProposalActionsService) {}
 
   @ResolveField(() => User)
   async user(@Parent() { userId }: ProposalActionEventHost) {
-    return this.proposalActionEventsService.getProposalActionEventHost(userId);
+    return this.proposalActionsService.getProposalActionEventHost(userId);
   }
 }
