+++ src/proposals/proposal-actions/proposal-action-group-configs.resolver.ts
@@ -1,6 +1,6 @@
 import { Parent, ResolveField, Resolver } from '@nestjs/graphql';
-import { ProposalAction } from '../models/proposal-action.model';
-import { ProposalActionsService } from '../proposal-actions.service';
+import { ProposalAction } from './models/proposal-action.model';
+import { ProposalActionsService } from './proposal-actions.service';
 import { ProposalActionGroupConfig } from './models/proposal-action-group-config.model';
 
 @Resolver(() => ProposalActionGroupConfig)
