+++ src/proposals/proposal-actions/proposal-action-role-members.resolver.ts
@@ -1,6 +1,6 @@
 import { Context, Parent, ResolveField, Resolver } from '@nestjs/graphql';
-import { Dataloaders } from '../../../dataloader/dataloader.types';
-import { User } from '../../../users/models/user.model';
+import { Dataloaders } from '../../dataloader/dataloader.types';
+import { User } from '../../users/models/user.model';
 import { ProposalActionRoleMember } from './models/proposal-action-role-member.model';
 
 @Resolver(() => ProposalActionRoleMember)
