+++ src/proposals/proposal-actions/proposal-actions.module.ts
@@ -1,27 +1,49 @@
 import { forwardRef, Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
+import { EventAttendee } from '../../events/event-attendees/models/event-attendee.model';
+import { Event } from '../../events/models/event.model';
 import { GroupConfigsModule } from '../../groups/group-configs/group-configs.module';
 import { GroupRolesModule } from '../../groups/group-roles/group-roles.module';
 import { Image } from '../../images/models/image.model';
 import { ProposalsModule } from '../proposals.module';
+import { ProposalActionEventHost } from './models/proposal-action-event-host.model';
+import { ProposalActionEvent } from './models/proposal-action-event.model';
+import { ProposalActionGroupConfig } from './models/proposal-action-group-config.model';
+import { ProposalActionPermission } from './models/proposal-action-permission.model';
+import { ProposalActionRoleMember } from './models/proposal-action-role-member.model';
+import { ProposalActionRole } from './models/proposal-action-role.model';
 import { ProposalAction } from './models/proposal-action.model';
-import { ProposalActionEventsModule } from './proposal-action-events/proposal-action-events.module';
-import { ProposalActionGroupConfigsModule } from './proposal-action-group-configs/proposal-action-group-configs.module';
-import { ProposalActionRolesModule } from './proposal-action-roles/proposal-action-roles.module';
 import { ProposalActionsResolver } from './proposal-actions.resolver';
 import { ProposalActionsService } from './proposal-actions.service';
+import { ProposalActionRolesResolver } from './proposal-action-roles.resolver';
+import { ProposalActionEventsResolver } from './proposal-action-events.resolver';
+import { ProposalActionEventHostsResolver } from './proposal-action-event-hosts.resolver';
 
 @Module({
   imports: [
-    TypeOrmModule.forFeature([ProposalAction, Image]),
-    forwardRef(() => ProposalActionEventsModule),
-    forwardRef(() => ProposalActionGroupConfigsModule),
+    TypeOrmModule.forFeature([
+      Event,
+      EventAttendee,
+      Image,
+      ProposalAction,
+      ProposalActionEvent,
+      ProposalActionEventHost,
+      ProposalActionGroupConfig,
+      ProposalActionPermission,
+      ProposalActionRole,
+      ProposalActionRoleMember,
+    ]),
     forwardRef(() => ProposalsModule),
     GroupConfigsModule,
     GroupRolesModule,
-    ProposalActionRolesModule,
   ],
-  providers: [ProposalActionsService, ProposalActionsResolver],
+  providers: [
+    ProposalActionEventHostsResolver,
+    ProposalActionEventsResolver,
+    ProposalActionRolesResolver,
+    ProposalActionsResolver,
+    ProposalActionsService,
+  ],
   exports: [ProposalActionsService],
 })
 export class ProposalActionsModule {}
