+++ src/proposals/proposal-actions/proposal-actions.resolver.ts
@@ -3,20 +3,14 @@ import { Image } from '../../images/models/image.model';
 import { Proposal } from '../models/proposal.model';
 import { ProposalsService } from '../proposals.service';
 import { ProposalAction } from './models/proposal-action.model';
-import { ProposalActionEvent } from './proposal-action-events/models/proposal-action-event.model';
-import { ProposalActionEventsService } from './proposal-action-events/proposal-action-events.service';
-import { ProposalActionRole } from './proposal-action-roles/models/proposal-action-role.model';
-import { ProposalActionRolesService } from './proposal-action-roles/proposal-action-roles.service';
+import { ProposalActionEvent } from './models/proposal-action-event.model';
+import { ProposalActionRole } from './models/proposal-action-role.model';
 import { ProposalActionsService } from './proposal-actions.service';
-import { ProposalActionGroupConfig } from './proposal-action-group-configs/models/proposal-action-group-config.model';
-import { ProposalActionGroupConfigsService } from './proposal-action-group-configs/proposal-action-group-configs.service';
+import { ProposalActionGroupConfig } from './models/proposal-action-group-config.model';
 
 @Resolver(() => ProposalAction)
 export class ProposalActionsResolver {
   constructor(
-    private proposalActionEventsService: ProposalActionEventsService,
-    private proposalActionGroupConfigsService: ProposalActionGroupConfigsService,
-    private proposalActionRolesService: ProposalActionRolesService,
     private proposalActionsService: ProposalActionsService,
     private proposalsService: ProposalsService,
   ) {}
@@ -28,21 +22,21 @@ export class ProposalActionsResolver {
 
   @ResolveField(() => ProposalActionRole)
   async role(@Parent() { id }: ProposalAction) {
-    return this.proposalActionRolesService.getProposalActionRole({
+    return this.proposalActionsService.getProposalActionRole({
       proposalActionId: id,
     });
   }
 
   @ResolveField(() => ProposalActionEvent)
   async event(@Parent() { id }: ProposalAction) {
-    return this.proposalActionEventsService.getProposalActionEvent({
+    return this.proposalActionsService.getProposalActionEvent({
       proposalActionId: id,
     });
   }
 
   @ResolveField(() => ProposalActionGroupConfig, { nullable: true })
   async groupSettings(@Parent() { id }: ProposalActionGroupConfig) {
-    return this.proposalActionGroupConfigsService.getProposalActionGroupConfig({
+    return this.proposalActionsService.getProposalActionGroupConfig({
       proposalActionId: id,
     });
   }
