+++ src/shield/rules/proposal.rules.ts
@@ -7,12 +7,12 @@ import { Image } from '../../images/models/image.model';
 import { ProposalConfig } from '../../proposals/models/proposal-config.model';
 import { Proposal } from '../../proposals/models/proposal.model';
 import { ProposalAction } from '../../proposals/proposal-actions/models/proposal-action.model';
-import { ProposalActionEventHost } from '../../proposals/proposal-actions/proposal-action-events/models/proposal-action-event-host.model';
-import { ProposalActionEvent } from '../../proposals/proposal-actions/proposal-action-events/models/proposal-action-event.model';
-import { ProposalActionGroupConfig } from '../../proposals/proposal-actions/proposal-action-group-configs/models/proposal-action-group-config.model';
-import { ProposalActionPermission } from '../../proposals/proposal-actions/proposal-action-roles/models/proposal-action-permission.model';
-import { ProposalActionRoleMember } from '../../proposals/proposal-actions/proposal-action-roles/models/proposal-action-role-member.model';
-import { ProposalActionRole } from '../../proposals/proposal-actions/proposal-action-roles/models/proposal-action-role.model';
+import { ProposalActionEventHost } from '../../proposals/proposal-actions/models/proposal-action-event-host.model';
+import { ProposalActionEvent } from '../../proposals/proposal-actions/models/proposal-action-event.model';
+import { ProposalActionGroupConfig } from '../../proposals/proposal-actions/models/proposal-action-group-config.model';
+import { ProposalActionPermission } from '../../proposals/proposal-actions/models/proposal-action-permission.model';
+import { ProposalActionRoleMember } from '../../proposals/proposal-actions/models/proposal-action-role-member.model';
+import { ProposalActionRole } from '../../proposals/proposal-actions/models/proposal-action-role.model';
 import { Vote } from '../../votes/models/vote.model';
 
 export const isOwnProposal = rule({ cache: 'strict' })(async (
@@ -79,14 +79,7 @@ export const isPublicProposalAction = rule({ cache: 'strict' })(async (
     | ProposalActionRole
     | ProposalActionRoleMember,
   _args,
-  {
-    services: {
-      proposalActionEventsService,
-      proposalActionRolesService,
-      proposalActionsService,
-      proposalsService,
-    },
-  }: Context,
+  { services: { proposalActionsService, proposalsService } }: Context,
 ) => {
   if (
     parent instanceof ProposalActionRole ||
@@ -106,7 +99,7 @@ export const isPublicProposalAction = rule({ cache: 'strict' })(async (
     parent instanceof ProposalActionPermission
   ) {
     const proposalActionRole =
-      await proposalActionRolesService.getProposalActionRole(
+      await proposalActionsService.getProposalActionRole(
         { id: parent.proposalActionRoleId },
         ['proposalAction.proposal.group.config'],
       );
@@ -117,7 +110,7 @@ export const isPublicProposalAction = rule({ cache: 'strict' })(async (
   }
   if (parent instanceof ProposalActionEventHost) {
     const proposalActionEvent =
-      await proposalActionEventsService.getProposalActionEvent(
+      await proposalActionsService.getProposalActionEvent(
         { id: parent.proposalActionEventId },
         ['proposalAction.proposal.group.config'],
       );
