+++ src/users/models/user.model.ts
@@ -20,7 +20,7 @@ import { Image } from '../../images/models/image.model';
 import { Like } from '../../likes/models/like.model';
 import { Post } from '../../posts/models/post.model';
 import { Proposal } from '../../proposals/models/proposal.model';
-import { ProposalActionRoleMember } from '../../proposals/proposal-actions/proposal-action-roles/models/proposal-action-role-member.model';
+import { ProposalActionRoleMember } from '../../proposals/proposal-actions/models/proposal-action-role-member.model';
 import { ServerInvite } from '../../server-invites/models/server-invite.model';
 import { ServerRole } from '../../server-roles/models/server-role.model';
